Processing extracts as lat lon on a single subject
This code will take in PRN data, make config files based on the workflow, extract the point data, convert the point data to lat-lon on a single subject, and cluster the resulting data.  `eps` and `min_samples` are left as variables at the top that can be adjusted to reasonable values (they are not at those values now).